Pre-processing data + BERT model for country label
Added the pre-processing and labeled the authors countries with BERT encoder, where string matching was not enough. 
Now we have countries where each author researched in. 
We include full names of at the authors but also abbreviated. 
The year and rest of columns are the same as in the non processed version + dropped the useless information. 

TO DO:

Maybe pre process the abstracts to get them ready for emmedigs. 
Use NER for country normalisation so we can keep the 85 entries labeled as unwon.